requirements.txt
DrissionPage
pandas
python-dotenv
requests
clickhouse-connect
openai